[UPDATE] upload and display images
