+++ server.js
@@ -4,19 +4,9 @@ const cors   = require("cors");
 const app = express();
 const db = require("./app/models");
 const PORT = process.env.PORT || 5000;
-const multer = require('multer');
 
 
-var storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, 'uploads')
-  },
-  filename: function (req, file, cb) {
-    cb(null, file.originalname )
-  }
-})
 
-var upload = multer({ storage: storage })
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
@@ -39,18 +29,9 @@ app.get("/", (req, res) => {
   res.sendFile(__dirname + '/upload.html');
 });
 
-app.post('/uploadfile', upload.single('myFile'), (req, res, next) => {
-  const file = req.file
-  if (!file) {
-    const error = new Error('Please upload a file')
-    error.httpStatusCode = 400
-    return next(error)
-  }
-    res.send(file)
- 
-})
 
 
+app.use(express.static(process.env.IMAGE_PATH));
 
 
 app.get("/test", (req, res) => {
@@ -68,6 +49,5 @@ require("./app/routes/orders.routes")(app);
 require("./app/routes/cats.routes")(app);
 require("./app/routes/owns.routes")(app);
 require("./app/routes/takes.routes")(app);
-
